# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReflectIQ API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FollowupInput {
  /** Student's answer to the open prompt (step 1) */
  step1Text: string;
  /** Selected course topic */
  topic: string;
}

export interface FollowupOutput {
  /** Generated clarifying follow-up question for step 2 */
  question: string;
}

export interface ReflectionInput {
  /** Selected course topic */
  topic: string;
  /** Answer to open prompt */
  step1Text: string;
  /** Answer to clarifying follow-up */
  step2Text: string;
  /** Answer to Socratic reasoning probe */
  step3Text: string;
}

export interface ReflectionResult {
  id: number;
  /** Classified learning-signal category */
  signal: string;
  /** Classification confidence 0-1 */
  confidence: number;
  /** @nullable */
  clusterId?: number | null;
}

export interface DashboardSummary {
  totalReflectionsThisWeek: number;
  /** Number of clusters meeting MIN_CLUSTER_SIZE threshold */
  activeClusterCount: number;
  /** Topics with high definitional_gap or causal_reasoning_gap concentration */
  flaggedTopicCount: number;
}

/**
 * Map of signal category to count
 */
export type HeatmapRowSignals = {[key: string]: number};

export interface HeatmapRow {
  topic: string;
  /** Map of signal category to count */
  signals: HeatmapRowSignals;
  total: number;
  /** True if this topic has high gap concentration */
  isHotspot: boolean;
}

export interface ClusterCard {
  id: number;
  topic: string;
  /** Learning signal category */
  signal: string;
  studentCount: number;
  /** 2-3 anonymized paraphrased phrases */
  representativePhrases: string[];
  /** One of: Revisit, Reinforce, Provide real-world examples, Clarify terminology, No action needed */
  suggestedAction: string;
  /**
     * null, acknowledged, or dismissed
     * @nullable
     */
  actionStatus: string | null;
  /** up, down, or stable */
  trendDirection: string;
  /**
     * Change in student count vs previous period
     * @nullable
     */
  trendDelta?: number | null;
}

export interface ClusterActionUpdate {
  /** acknowledged or dismissed */
  actionStatus: string;
}

export interface SeedResult {
  reflectionsCreated: number;
  clustersCreated: number;
  /** Clusters below MIN_CLUSTER_SIZE that were created in raw data but suppressed */
  suppressedClusters: number;
}

export interface RawStats {
  totalRawReflections: number;
  visibleClusters: number;
  suppressedClusters: number;
  minClusterSize: number;
}

export type GetHeatmapParams = {
/**
 * Filter by course topic
 */
topic?: string;
/**
 * ISO date string for start of date range
 */
dateFrom?: string;
/**
 * ISO date string for end of date range
 */
dateTo?: string;
/**
 * Filter by signal type
 */
signal?: string;
};

export type GetClustersParams = {
/**
 * Filter by course topic
 */
topic?: string;
/**
 * ISO date string for start of date range
 */
dateFrom?: string;
/**
 * ISO date string for end of date range
 */
dateTo?: string;
/**
 * Filter by signal type
 */
signal?: string;
};