# lib/api-zod/src/generated/types/clusterActionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReflectIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClusterActionUpdate {
  /** acknowledged or dismissed */
  actionStatus: string;
}